+++ src/scaup/crud/top_level_containers.py
@@ -1,3 +1,5 @@
+from typing import Any
+
 from fastapi import HTTPException, status
 from lims_utils.logging import app_logger
 from lims_utils.models import Paged
@@ -19,6 +21,17 @@ from ..utils.session import retry_if_exists
 DEWAR_PREFIX = "DLS-BI-1"
 
 
+def _check_if_dls_code(code: str | None):
+    """Check if code is DLS barcode, or user provided serial number
+
+    Args:
+        code: User-provided code
+
+    Returns:
+        True if it is a DLS code, False otherwise"""
+    return code is not None and code[:3] == "DLS"
+
+
 def _check_fields(
     params: TopLevelContainerIn | OptionalTopLevelContainer,
     token: str,
@@ -36,7 +49,6 @@ def _check_fields(
         if params.code is None:
             # Perform no facility code check if code is not present
             return
-
         query = query.select_from(TopLevelContainer).filter(TopLevelContainer.id == item_id).join(Shipment)
 
     proposal_reference = inner_db.session.scalar(query)
@@ -70,7 +82,7 @@ def create_top_level_container(shipmentId: int | None, params: TopLevelContainer
         ).one()
     )
 
-    if params.code:
+    if _check_if_dls_code(params.code):
         _check_fields(params, token, shipmentId)
     elif params.type == "dewar" and autocreate:
         # Automatically register dewar if no code is provided
@@ -100,7 +112,10 @@ def create_top_level_container(shipmentId: int | None, params: TopLevelContainer
                 Config.ispyb_api.jwt,
                 method="POST",
                 url=f"/proposals/{proposal.reference}/dewar-registry",
-                json={"facilityCode": new_code},
+                json={
+                    "facilityCode": new_code,
+                    "manufacturerSerialNumber": (None if params.code.lower().strip() == "n/a" else params.code),
+                },
             )
 
             if ext_resp.status_code != 201:
@@ -125,7 +140,29 @@ def create_top_level_container(shipmentId: int | None, params: TopLevelContainer
     )
 
     if proposal:
-        bar_code = f"{proposal.reference}-{proposal.visitNumber}-{container.id:07}"
+        # This is required because the dewar logistics server expects an instrument in the barcode in order to match
+        # a dewar to the correct instrument
+        ext_resp = ExternalRequest.request(
+            token=token,
+            url=f"/proposals/{proposal.reference}/sessions/{proposal.visitNumber}",
+        )
+
+        if ext_resp.status_code != 200:
+            app_logger.warning(
+                "Error from Expeye while getting session %s-%i: %s",
+                proposal.reference,
+                proposal.visitNumber,
+                ext_resp.text,
+            )
+            raise HTTPException(
+                status_code=status.HTTP_424_FAILED_DEPENDENCY,
+                detail="Invalid response while creating top level container in ISPyB",
+            )
+
+        session_json: dict[str, Any] = ext_resp.json()
+        instrument = "-" if (i := session_json.get("beamLineName")) is None else f"-{i}-"
+
+        bar_code = f"{proposal.reference}-{proposal.visitNumber}{instrument}{container.id:07}"
 
         # This is because some users expect a sequential numeric ID to make tracking how old a dewar is easier,
         # and because of historical reasons, some users are used to seeing the proposal/session number on there