+++ tests/shipments/top_level_containers/test_create.py
@@ -1,3 +1,5 @@
+import json
+
 import pytest
 import responses
 from sqlalchemy import select
@@ -14,7 +16,7 @@ def test_create_invalid_code(client):
         "/shipments/1/topLevelContainers",
         json={
             "type": "dewar",
-            "code": "DOESNOTEXIST",
+            "code": "DLS-DOES-NOT-EXIST",
         },
     )
 
@@ -41,8 +43,8 @@ def test_create_no_name(client):
 
 
 @responses.activate
-def test_create_auto_barcode(client):
-    """Should automatically generate barcode if not provided in request"""
+def test_create_auto_barcode_no_instrument(client):
+    """Should automatically generate barcode if not provided in request, and instrument isn't returned from ISPyB"""
 
     resp = client.post(
         "/shipments/1/topLevelContainers",
@@ -58,6 +60,60 @@ def test_create_auto_barcode(client):
     assert "cm1-1-" in inner_db.session.scalar(
         select(TopLevelContainer.barCode).filter(TopLevelContainer.id == new_tlc)
     )
+
+
+@pytest.mark.noregister
+@responses.activate
+def test_create_auto_barcode_with_instrument(client):
+    """Should automatically generate barcode if not provided in request"""
+    responses.get(
+        f"{Config.ispyb_api.url}/proposals/cm1/sessions/1",
+        status=200,
+        json={"beamLineName": "m02"},
+    )
+
+    responses.get(
+        f"{Config.ispyb_api.url}/proposals/cm1/dewar-registry/DLS-EM-0001",
+        status=200,
+        json={},
+    )
+
+    resp = client.post(
+        "/shipments/1/topLevelContainers",
+        json={
+            "type": "dewar",
+            "code": "DLS-EM-0001",
+        },
+    )
+
+    assert resp.status_code == 201
+    new_tlc = resp.json()["id"]
+    assert "cm1-1-m02-" in inner_db.session.scalar(
+        select(TopLevelContainer.barCode).filter(TopLevelContainer.id == new_tlc)
+    )
+
+
+@pytest.mark.noregister
+@responses.activate
+def test_create_auto_barcode_upstream_fail(client):
+    """Should raise exception if request fails upstream when generating barcode"""
+    responses.get(f"{Config.ispyb_api.url}/proposals/cm1/sessions/1", status=500)
+
+    responses.get(
+        f"{Config.ispyb_api.url}/proposals/cm1/dewar-registry/DLS-EM-0001",
+        status=200,
+        json={},
+    )
+
+    resp = client.post(
+        "/shipments/1/topLevelContainers",
+        json={
+            "type": "dewar",
+            "code": "DLS-EM-0001",
+        },
+    )
+
+    assert resp.status_code == 424
 
 
 @responses.activate
@@ -104,6 +160,41 @@ def test_create_no_code(client):
     assert (
         inner_db.session.scalar(select(TopLevelContainer).filter(TopLevelContainer.code == "DLS-BI-5672")) is not None
     )
+
+
+@responses.activate
+def test_create_with_serial(client):
+    """Should automatically generate code if serial code is provided"""
+    creation_response = responses.post(
+        f"{Config.ispyb_api.url}/proposals/cm1/dewar-registry",
+        status=201,
+        json={"dewarRegistryId": 1},
+    )
+
+    responses.get(
+        f"{Config.ispyb_api.url}/dewar-registry?search=DLS-BI-1&limit=1",
+        status=200,
+        json={"items": [{"facilityCode": "DLS-BI-5671"}]},
+    )
+
+    resp = client.post(
+        "/shipments/1/topLevelContainers",
+        json={
+            "type": "dewar",
+            "code": "test123",
+        },
+    )
+
+    assert resp.status_code == 201
+
+    assert (
+        inner_db.session.scalar(select(TopLevelContainer).filter(TopLevelContainer.code == "DLS-BI-5672")) is not None
+    )
+
+    assert json.loads(creation_response.calls[0].request.body.decode()) == {
+        "facilityCode": "DLS-BI-5672",
+        "manufacturerSerialNumber": "test123",
+    }
 
 
 @responses.activate